+++ src/sumologic-span-processor/document-visibility-state.ts
@@ -6,24 +6,27 @@ import {
 } from '@opentelemetry/sdk-trace-base';
 
 const ATTRIBUTE_NAME = 'document.visibilityState';
-const VISIBILITY_STATE_TO_EVENT_NAMES: Record<VisibilityState, string> = {
-  visible: 'pageshow',
-  hidden: 'pagehide',
-};
+const VISIBILITY_STATE_TO_EVENT_NAMES: Record<DocumentVisibilityState, string> =
+  {
+    visible: 'pageshow',
+    hidden: 'pagehide',
+  };
 const MAX_CHANGES = 100;
 
 const changes: {
   timestampInNanoseconds: number;
   timestampInHrTime: HrTime;
-  state: VisibilityState;
+  state: DocumentVisibilityState;
 }[] = [];
-let currentState = document.visibilityState;
+let currentState: DocumentVisibilityState | undefined;
 
 // exported for tests
 export const resetDocumentVisibilityStateChanges = () => {
   while (changes.length) {
     changes.pop();
   }
+  currentState = undefined;
+  updateState();
 };
 
 const updateState = () => {
@@ -42,6 +45,8 @@ const updateState = () => {
   }
 };
 
+updateState(); // we want to keep the initial state in the history as well
+
 document.addEventListener('visibilitychange', () => {
   updateState();
 });
@@ -55,7 +60,14 @@ window.addEventListener('pageshow', () => {
 });
 
 export const onStart = (span: SdkTraceSpan, context?: Context): void => {
-  span.setAttribute(ATTRIBUTE_NAME, currentState);
+  const startTimeInNanoseconds = hrTimeToNanoseconds(span.startTime);
+  for (let i = changes.length - 1; i >= 0; i -= 1) {
+    const { timestampInNanoseconds, state } = changes[i];
+    if (timestampInNanoseconds <= startTimeInNanoseconds) {
+      span.setAttribute(ATTRIBUTE_NAME, state);
+      break;
+    }
+  }
 };
 
 export const onEnd = (readableSpan: ReadableSpan): void => {
@@ -69,7 +81,9 @@ export const onEnd = (readableSpan: ReadableSpan): void => {
     ? hrTimeToNanoseconds(span.endTime)
     : Infinity;
 
-  for (let i = changes.length - 1; i >= 0; i -= 1) {
+  // we skip the initial change because it's already covered by the onStart function
+  // and we don't want to save initial state as an event
+  for (let i = changes.length - 1; i >= 1; i -= 1) {
     const { timestampInNanoseconds, timestampInHrTime, state } = changes[i];
     if (timestampInNanoseconds < startTimeInNanoseconds) {
       break;
@@ -83,7 +97,7 @@ export const onEnd = (readableSpan: ReadableSpan): void => {
         attributes: undefined,
         time: timestampInHrTime,
       });
-      if (state === 'hidden' && span.attributes[ATTRIBUTE_NAME] === 'visible') {
+      if (state === 'hidden') {
         span.attributes[ATTRIBUTE_NAME] = state;
       }
     }